// src/app/shared/config.interface.ts
export class Config {
    age: number;
    gender: string;
    height: number;
    weight: number;
    systolicBloodPressure: number;
    diastolicBloodPressure: number;
    bloodGlucose: number;
    bloodOxygen: number;
    tobaccoUse: string;
    lungSound: string;
    quantity: number;
    saveToFile: boolean;
    watching: boolean;
    typeOfGenerating: string;
    dateAndTime: string;
    endPoints: string[];
}